Implement drive mode switching
Users can now switch between tank and split-stick arcade by
pressing triangle button on controller.

Added new command class SwtichDriveModes as an InstantCommand,
however may be overkill for this (simply calling a function
from the drive subsystem to do the switching).

Added inline code to do the saem thing in RobotContainer. The use of
the SwitchDriveModes command is commented out. Left there as an
example, and for testing.

Updated configureBindings() in RobotContainer, with the command
for switching.

Updated DriveSubsystem with a member function for switching drive modes.